Splitter filer og flytter typer inn i pakker
