+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/tidslinje/Periode.kt
@@ -1,5 +1,8 @@
 package no.nav.familie.ba.sak.kjerne.tidslinje
 
+import no.nav.familie.ba.sak.kjerne.tidslinje.tid.Tidspunkt
+import no.nav.familie.ba.sak.kjerne.tidslinje.tid.Tidsrom
+
 data class Periode<T>(
     val fom: Tidspunkt,
     val tom: Tidspunkt,
