+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/tidslinje/Tidslinje.kt
@@ -1,5 +1,7 @@
 package no.nav.familie.ba.sak.kjerne.tidslinje
 
+import no.nav.familie.ba.sak.kjerne.tidslinje.tid.Tidsrom
+
 abstract class Tidslinje<T> {
     internal abstract fun tidsrom(): Tidsrom
     abstract fun perioder(): Collection<Periode<T>>
@@ -51,65 +53,3 @@ abstract class Tidslinje<T> {
         }
     }
 }
-
-abstract class TidslinjeMedAvhengigheter<T>(
-    private val foregåendeTidslinjer: Collection<Tidslinje<*>>
-) : Tidslinje<T>() {
-
-    init {
-        if (foregåendeTidslinjer.size == 0)
-            throw IllegalStateException("Skal ha avhengigheter, men listen over avhengigher er tom")
-    }
-
-    override fun tidsrom(): Tidsrom {
-        val fom = foregåendeTidslinjer.map { it.tidsrom().start }.minsteEllerUendelig()
-        val tom = foregåendeTidslinjer.map { it.tidsrom().endInclusive }.størsteEllerUendelig()
-
-        return fom..tom
-    }
-}
-
-abstract class SnittTidslinje<T>(
-    avhengigheter: Collection<Tidslinje<*>>,
-) : TidslinjeMedAvhengigheter<T>(avhengigheter) {
-
-    var generertePerioder: Collection<Periode<T>>? = null
-
-    constructor(vararg avhengighet: Tidslinje<*>) :
-        this(avhengighet.asList())
-
-    override fun perioder(): Collection<Periode<T>> {
-        return generertePerioder ?: genererPerioder().also { generertePerioder = it }
-    }
-
-    private fun genererPerioder(): List<Periode<T>> =
-        tidsrom().map { Pair(it, beregnSnitt(it)) }
-            .fold(emptyList()) { acc, (tidspunkt, innhold) ->
-                val sistePeriode = acc.lastOrNull()
-                when {
-                    sistePeriode != null && sistePeriode.innhold == innhold && sistePeriode.tom.erRettFør(tidspunkt.somEndelig()) ->
-                        acc.replaceLast(sistePeriode.copy(tom = tidspunkt.somTilOgMed()))
-                    else -> acc + Periode(tidspunkt.somFraOgMed(), tidspunkt.somTilOgMed(), innhold)
-                }
-            }
-
-    protected abstract fun beregnSnitt(tidspunkt: Tidspunkt): T?
-
-    companion object {
-        private fun <T> Collection<T>.replaceLast(replacement: T) =
-            this.take(this.size - 1) + replacement
-    }
-}
-
-fun <T> Tidslinje<T>.hentUtsnitt(tidspunkt: Tidspunkt): T? =
-    perioder().hentUtsnitt(tidspunkt)
-
-fun <T> Collection<Periode<T>>.hentUtsnitt(tidspunkt: Tidspunkt): T? =
-    this.filter { it.fom <= tidspunkt && it.tom >= tidspunkt }
-        .firstOrNull()?.innhold
-
-class KomprimerendeTidslinje<T>(val tidslinje: Tidslinje<T>) : SnittTidslinje<T>(tidslinje) {
-    override fun beregnSnitt(tidspunkt: Tidspunkt): T? = tidslinje.hentUtsnitt(tidspunkt)
-}
-
-fun <T> Tidslinje<T>.komprimer(): Tidslinje<T> = KomprimerendeTidslinje(this)
