--- src/main/kotlin/no/nav/familie/ba/sak/kjerne/tidslinje/Tidspunkt.kt
@@ -1,294 +1,0 @@
-package no.nav.familie.ba.sak.kjerne.tidslinje
-
-import no.nav.familie.ba.sak.common.sisteDagIMåned
-import no.nav.familie.ba.sak.common.toYearMonth
-import java.time.LocalDate
-import java.time.YearMonth
-
-val PRAKTISK_SENESTE_DAG = LocalDate.of(2499, 12, 31)
-val PRAKTISK_TIDLIGSTE_DAG = LocalDate.of(1900, 1, 1)
-
-private val månedTilDagKonverterer: (YearMonth) -> LocalDate = { it.atEndOfMonth() }
-private val dagTilMånedKonverterer: (LocalDate) -> YearMonth = { it.toYearMonth() }
-
-enum class Uendelighet {
-    INGEN,
-    FORTID,
-    FREMTID
-}
-
-interface Tidspunkt : Comparable<Tidspunkt> {
-    val uendelighet: Uendelighet
-
-    fun tilFørsteDagIMåneden(): DagTidspunkt
-    fun tilSisteDagIMåneden(): DagTidspunkt
-    fun tilDagIMåned(dag: Int): DagTidspunkt
-    fun tilDag(månedTilDagMapper: (YearMonth) -> LocalDate): DagTidspunkt
-    fun tilInneværendeMåned(): MånedTidspunkt
-    fun tilLocalDateEllerNull(): LocalDate?
-    fun tilLocalDate(): LocalDate
-    fun tilYearMonthEllerNull(): YearMonth?
-    fun tilYearMonth(): YearMonth
-
-    fun flytt(tidsenheter: Long): Tidspunkt
-    fun neste() = flytt(1)
-    fun forrige() = flytt(-1)
-
-    fun erRettFør(tidspunkt: Tidspunkt) = neste() == tidspunkt
-    fun erEndelig(): Boolean = uendelighet == Uendelighet.INGEN
-    fun erUendeligLengeSiden(): Boolean = uendelighet == Uendelighet.FORTID
-    fun erUendeligLengeTil(): Boolean = uendelighet == Uendelighet.FREMTID
-
-    fun somEndelig(): Tidspunkt
-    fun somUendeligLengeSiden(): Tidspunkt
-    fun somUendeligLengeTil(): Tidspunkt
-    fun somFraOgMed(): Tidspunkt
-    fun somTilOgMed(): Tidspunkt
-
-    fun erDag(): Boolean
-    fun erMåned() = !erDag()
-
-    companion object {
-        fun uendeligLengeSiden(dato: LocalDate) = DagTidspunkt(dato, uendelighet = Uendelighet.FORTID)
-        fun uendeligLengeSiden(måned: YearMonth) = MånedTidspunkt(måned, Uendelighet.FORTID)
-        fun uendeligLengeTil(dato: LocalDate) = DagTidspunkt(dato, uendelighet = Uendelighet.FREMTID)
-        fun uendeligLengeTil(måned: YearMonth) = MånedTidspunkt(måned, Uendelighet.FREMTID)
-        fun fraOgMed(fraOgMed: LocalDate?, praktiskMinsteFraOgMed: LocalDate) =
-            if (fraOgMed == null || fraOgMed < PRAKTISK_TIDLIGSTE_DAG)
-                uendeligLengeSiden(maxOf(praktiskMinsteFraOgMed, PRAKTISK_TIDLIGSTE_DAG))
-            else
-                DagTidspunkt(fraOgMed, Uendelighet.INGEN)
-
-        fun tilOgMed(tilOgMed: LocalDate?, praktiskStørsteTilOgMed: LocalDate) =
-            if (tilOgMed == null || tilOgMed > PRAKTISK_SENESTE_DAG)
-                uendeligLengeTil(minOf(praktiskStørsteTilOgMed, PRAKTISK_SENESTE_DAG))
-            else
-                DagTidspunkt(tilOgMed, Uendelighet.INGEN)
-
-        fun med(dato: LocalDate) = DagTidspunkt(dato, Uendelighet.INGEN)
-        fun med(måned: YearMonth) = MånedTidspunkt(måned, Uendelighet.INGEN)
-    }
-
-    // Betrakter to uendeligheter som like, selv underliggende tidspunkt kan være forskjellig
-    override fun compareTo(other: Tidspunkt) =
-        if (this.uendelighet == Uendelighet.FORTID && other.uendelighet == Uendelighet.FORTID)
-            0
-        else if (this.uendelighet == Uendelighet.FREMTID && other.uendelighet == Uendelighet.FREMTID)
-            0
-        else if (this.uendelighet == Uendelighet.FORTID && other.uendelighet != Uendelighet.FORTID)
-            -1
-        else if (this.uendelighet == Uendelighet.FREMTID && other.uendelighet != Uendelighet.FREMTID)
-            1
-        else if (this.uendelighet != Uendelighet.FORTID && other.uendelighet == Uendelighet.FORTID)
-            1
-        else if (this.uendelighet != Uendelighet.FREMTID && other.uendelighet == Uendelighet.FREMTID)
-            -1
-        else
-            this.tilLocalDate().compareTo(other.tilLocalDate())
-}
-
-data class DagTidspunkt(
-    private val dato: LocalDate,
-    override val uendelighet: Uendelighet
-) : Tidspunkt {
-
-    init {
-        if (dato < PRAKTISK_TIDLIGSTE_DAG)
-            throw IllegalArgumentException("Kan ikke håndtere så tidlig tidspunkt. Bruk uendeligLengeSiden()")
-        else if (dato > PRAKTISK_SENESTE_DAG)
-            throw IllegalArgumentException("Kan ikke håndtere så sent tidspunkt. Bruk uendeligLengeTil()")
-    }
-
-    override fun tilFørsteDagIMåneden(): DagTidspunkt {
-        return this.copy(dato = dato.withDayOfMonth(1))
-    }
-
-    override fun tilSisteDagIMåneden(): DagTidspunkt {
-        return this.copy(dato = dato.sisteDagIMåned())
-    }
-
-    override fun tilDagIMåned(dag: Int): DagTidspunkt {
-        return this.copy(dato = dato.withDayOfMonth(dag))
-    }
-
-    override fun tilDag(månedTilDagMapper: (YearMonth) -> LocalDate): DagTidspunkt = this
-
-    override fun tilInneværendeMåned(): MånedTidspunkt {
-        return MånedTidspunkt(dagTilMånedKonverterer(this.dato), uendelighet)
-    }
-
-    override fun tilLocalDateEllerNull(): LocalDate? {
-        if (uendelighet != Uendelighet.INGEN)
-            return null
-        else
-            return dato
-    }
-
-    override fun tilLocalDate(): LocalDate {
-        return tilLocalDateEllerNull() ?: throw IllegalStateException("Tidspunkt er uendelig")
-    }
-
-    override fun tilYearMonthEllerNull(): YearMonth? {
-        return tilLocalDateEllerNull()?.let { dagTilMånedKonverterer(it) }
-    }
-
-    override fun tilYearMonth(): YearMonth {
-        return dagTilMånedKonverterer(tilLocalDate())
-    }
-
-    override fun flytt(tidsenheter: Long): Tidspunkt {
-        return this.copy(dato = dato.plusDays(tidsenheter))
-    }
-
-    override fun somEndelig(): Tidspunkt {
-        return copy(uendelighet = Uendelighet.INGEN)
-    }
-
-    override fun somUendeligLengeSiden(): Tidspunkt {
-        return copy(uendelighet = Uendelighet.FORTID)
-    }
-
-    override fun somUendeligLengeTil(): Tidspunkt {
-        return copy(uendelighet = Uendelighet.FREMTID)
-    }
-
-    override fun somFraOgMed(): Tidspunkt {
-        return if (uendelighet == Uendelighet.FREMTID)
-            somEndelig()
-        else
-            this
-    }
-
-    override fun somTilOgMed(): Tidspunkt {
-        return if (uendelighet == Uendelighet.FORTID)
-            somEndelig()
-        else
-            this
-    }
-
-    override fun erDag() = true
-    override fun toString(): String {
-        return when (uendelighet) {
-            Uendelighet.FORTID -> "<--"
-            else -> ""
-        } + dato + when (uendelighet) {
-            Uendelighet.FREMTID -> "-->"
-            else -> ""
-        }
-    }
-}
-
-data class MånedTidspunkt(
-    private val måned: YearMonth,
-    override val uendelighet: Uendelighet
-) : Tidspunkt {
-    init {
-        if (måned < PRAKTISK_TIDLIGSTE_DAG.toYearMonth())
-            throw IllegalArgumentException("Tidspunktet er for lenge siden. Bruk uendeligLengeSiden()")
-        else if (måned > PRAKTISK_SENESTE_DAG.toYearMonth())
-            throw IllegalArgumentException("Tidspunktet er for lenge til. Bruk uendeligLengeTil()")
-    }
-
-    override fun tilFørsteDagIMåneden(): DagTidspunkt =
-        DagTidspunkt(måned.atDay(1), this.uendelighet)
-
-    override fun tilSisteDagIMåneden(): DagTidspunkt =
-        DagTidspunkt(måned.atEndOfMonth(), this.uendelighet)
-
-    override fun tilDagIMåned(dag: Int): DagTidspunkt =
-        DagTidspunkt(måned.atDay(dag), this.uendelighet)
-
-    override fun tilDag(månedTilDagMapper: (YearMonth) -> LocalDate): DagTidspunkt =
-        DagTidspunkt(månedTilDagMapper(måned), uendelighet = uendelighet)
-
-    override fun tilInneværendeMåned() = this
-
-    override fun tilLocalDateEllerNull(): LocalDate? =
-        tilYearMonthEllerNull()?.let(månedTilDagKonverterer)
-
-    override fun tilLocalDate(): LocalDate = månedTilDagKonverterer(måned)
-
-    override fun tilYearMonthEllerNull(): YearMonth? =
-        if (uendelighet != Uendelighet.INGEN)
-            null
-        else
-            måned
-
-    override fun tilYearMonth(): YearMonth =
-        if (uendelighet != Uendelighet.INGEN)
-            throw IllegalStateException("Tidspunktet er uendelig")
-        else
-            måned
-
-    override fun flytt(tidsenheter: Long) = copy(måned = måned.plusMonths(tidsenheter))
-
-    override fun somEndelig(): Tidspunkt = this.copy(uendelighet = Uendelighet.INGEN)
-
-    override fun toString(): String {
-        return when (uendelighet) {
-            Uendelighet.FORTID -> "<--"
-            else -> ""
-        } + måned + when (uendelighet) {
-            Uendelighet.FREMTID -> "-->"
-            else -> ""
-        }
-    }
-
-    override fun somUendeligLengeSiden(): Tidspunkt = this.copy(uendelighet = Uendelighet.FORTID)
-    override fun somUendeligLengeTil(): Tidspunkt = this.copy(uendelighet = Uendelighet.FREMTID)
-    override fun somFraOgMed(): Tidspunkt {
-        return if (uendelighet == Uendelighet.FREMTID) {
-            somEndelig()
-        } else
-            this
-    }
-
-    override fun somTilOgMed(): Tidspunkt {
-        return if (uendelighet == Uendelighet.FORTID) {
-            somEndelig()
-        } else
-            this
-    }
-
-    override fun erDag() = false
-}
-
-fun størsteAv(t1: Tidspunkt, t2: Tidspunkt): Tidspunkt =
-    if (t1.erUendeligLengeTil() && t2.erEndelig() && t1.somEndelig() <= t2)
-        t2.neste().somUendeligLengeTil()
-    else if (t2.erUendeligLengeTil() && t1.erEndelig() && t2.somEndelig() <= t1)
-        t1.neste().somUendeligLengeTil()
-    else if (t1.erUendeligLengeTil() || t2.erUendeligLengeTil())
-        maxOf(t1.somEndelig(), t2.somEndelig()).somUendeligLengeTil()
-    else
-        maxOf(t1, t2)
-
-fun minsteAv(t1: Tidspunkt, t2: Tidspunkt): Tidspunkt =
-    if (t1.erUendeligLengeSiden() && t2.erEndelig() && t1.somEndelig() >= t2)
-        t2.forrige().somUendeligLengeSiden()
-    else if (t2.erUendeligLengeSiden() && t1.erEndelig() && t2.somEndelig() >= t1)
-        t1.forrige().somUendeligLengeSiden()
-    else if (t1.erUendeligLengeSiden() || t2.erUendeligLengeSiden())
-        minOf(t1.somEndelig(), t2.somEndelig()).somUendeligLengeSiden()
-    else
-        minOf(t1, t2)
-
-fun Iterable<Tidspunkt>.størsteEllerUendelig() =
-    this.reduce { acc, neste ->
-        størsteAv(acc, neste)
-    }
-
-fun Iterable<Tidspunkt>.minsteEllerUendelig() =
-    this.reduce { acc, neste -> minsteAv(acc, neste) }
-
-fun LocalDate?.tilTidspunktEllerUendeligLengeSiden(default: () -> LocalDate) =
-    this?.let { DagTidspunkt(this, Uendelighet.INGEN) } ?: Tidspunkt.uendeligLengeSiden(default())
-
-fun LocalDate?.tilTidspunktEllerUendeligLengeTil(default: () -> LocalDate) =
-    this?.let { DagTidspunkt(this, Uendelighet.INGEN) } ?: Tidspunkt.uendeligLengeTil(default())
-
-fun YearMonth?.tilTidspunktEllerUendeligLengeSiden(default: () -> YearMonth) =
-    this?.let { MånedTidspunkt(this, Uendelighet.INGEN) } ?: Tidspunkt.uendeligLengeSiden(default())
-
-fun YearMonth?.tilTidspunktEllerUendeligLengeTil(default: () -> YearMonth) =
-    this?.let { MånedTidspunkt(this, Uendelighet.INGEN) } ?: Tidspunkt.uendeligLengeTil(default())
