+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/tidslinje/komposisjon/TidslinjeKombinator.kt
@@ -1,0 +1,9 @@
+package no.nav.familie.ba.sak.kjerne.tidslinje
+
+interface ToveisKombinator<V, H, R> {
+    fun kombiner(venstre: V?, høyre: H?): R
+}
+
+interface ListeKombinator<T, R> {
+    fun kombiner(liste: Iterable<T>): R
+}
