+++ src/test/kotlin/no/nav/familie/ba/sak/kjerne/tidslinje/util/CharTidslinje.kt
@@ -2,10 +2,10 @@ package no.nav.familie.ba.sak.kjerne.tidslinje.util
 
 import no.nav.familie.ba.sak.kjerne.tidslinje.Periode
 import no.nav.familie.ba.sak.kjerne.tidslinje.Tidslinje
-import no.nav.familie.ba.sak.kjerne.tidslinje.Tidspunkt
-import no.nav.familie.ba.sak.kjerne.tidslinje.Tidsrom
-import no.nav.familie.ba.sak.kjerne.tidslinje.komprimer
-import no.nav.familie.ba.sak.kjerne.tidslinje.rangeTo
+import no.nav.familie.ba.sak.kjerne.tidslinje.komposisjon.komprimer
+import no.nav.familie.ba.sak.kjerne.tidslinje.tid.Tidspunkt
+import no.nav.familie.ba.sak.kjerne.tidslinje.tid.Tidsrom
+import no.nav.familie.ba.sak.kjerne.tidslinje.tid.rangeTo
 import java.time.YearMonth
 
 internal class CharTidslinje(private val tegn: String, private val startMåned: Tidspunkt) : Tidslinje<Char>() {
